# openapi.yaml
...
openapi: "3.0.0"
info:
  title: "API References"
  description: "The **Linkage API** is built on RESTful principles and returns JSON\
    \ encoded-responses and accepts JSON payloads. It is only available through HTTPS\
    \ to ensure the security of the data in transfer.\n\nThe API operates in two environments:\
    \ **production** and **sandbox**.\n\nSandbox environment returns mock data and\
    \ is used for testing the basic fucntionality of the system.\n\nProduction environment\
    \ is used to retrieve the real accounts connected through Linkage.           \
    \  "
  contact:
    name: "Linkage"
    url: "https://linkagehq.com/contact-us"
    email: "contact@linkagehq.com"
  version: "v1.0"
servers:
- url: "https://api.linkagehq.com/v1/"
  description: "Production Server endpoint to connect data with Linkage."
- url: "https://api-sandbox.linkagehq.com/v1/"
  description: "sandbox server which will hepls you simulate the system for testing\
    \ purposes."
- url: "https://virtserver.swaggerhub.com/LinkageHQ/Api-Documentation/v1"
  description: "SwaggerHub API Auto Mocking"
...
paths:
...
  /users/{id}:
    get:
      tags:
      - "User"
      summary: "Retrieve a user"
      description: "Retrieves a user object with the supplied ID."
      operationId: "getUserById"
      parameters:
      - name: "id"
        in: "path"
        description: "The identifier of the user to be retrieved."
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Returns a user object if a valid identifier was provided."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetails"
  /users:
    get:
      tags:
      - "User"
      summary: "List users"
      description: "Lists all users that used the plugin configured with your plugin\
        \ key."
      operationId: "getAllUsers"
      parameters:
      - name: "limit"
        in: "query"
        description: "The number of user objects to be returned. The default is 10."
        required: false
        style: "form"
        explode: true
        schema:
          type: "integer"
      - name: "offset"
        in: "query"
        description: "The number of user objects to skip. Use this argument for pagination.\
          \ The default is 0."
        required: false
        style: "form"
        explode: true
        schema:
          type: "integer"
      responses:
        "200":
          description: "An object with a results property that contains an array of\
            \ up to limit user objects."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/UserDetails"
...
  /reputations/{id}:
    get:
      tags:
      - "Reputations"
      summary: "Retrieve a reputation"
      description: "Retrieves a reputation object with the supplied ID."
      operationId: "getReputationById"
      parameters:
      - name: "id"
        in: "path"
        description: "User ID"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReputationDetails"
...
  /vehicles/{id}:
    get:
      tags:
      - "Vehicles"
      summary: "Retrieve a vehicle"
      description: "Retrieves a vehicle object with the supplied ID."
      operationId: "getVehicleById"
      parameters:
      - name: "id"
        in: "path"
        description: "The identifier of the vehicle to be retrieved."
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Returns a vehicle object if a valid identifier was provided."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicles"
...
  /finances/{id}:
    get:
      tags:
      - "Finances"
      summary: "Retrieve a finance"
      description: "Retrieves a payout object with the supplied ID."
      operationId: "getFinaceById"
      parameters:
      - name: "id"
        in: "path"
        description: "The identifier of the payout to be retrieved."
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Returns a finance object if a valid identifier was provided."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payouts"
components:
  schemas:
...
    UserDetails:
      type: "object"
      properties:
        id:
          type: "string"
          description: "Object's unique identifier."
          example: "0994b847-cc69-4d98-bc99-c5e65e762add"
        created_at:
          type: "string"
          description: "Time at which the user was created. Timestamps follow the\
            \ ISO 8601 standard, for"
          example: "2019-11-27T15:56:50.583233Z"
        data_providers_connected:
          type: "array"
          description: "A list of connected data partner IDs"
          items:
            type: "string"
      description: "User objects represent distinct users that used linkagehq plugin\
        \ to connect their workforce accounts."
...
    ReputationDetails:
      type: "object"
      properties:
        id:
          type: "string"
          example: "47b216e2-d334-4235-bc1e-185d15ab18d0"
        client:
          type: "string"
          example: "010db8b4-a724-47fc-a17e-733b656312a2"
        created_by:
          type: "string"
          example: "82bc04fa-1e11-43f2-9a88-357f865b4998"
        created_at:
          type: "string"
          example: "2019-11-27T15:56:50.583233Z"
        updated_at:
          type: "string"
          example: "2019-11-27T15:56:50.583233Z"
        rating:
          type: "string"
          example: "4.95"
        acceptance_rate:
          type: "string"
          example: "0.86"
        ontime_rate:
          type: "string"
          example: "0.8"
        account:
          type: "string"
          example: "010db8b4-a724-47fc-a17e-733b656312a2"
        achievements:
          type: "array"
          items:
            type: "object"
            properties:
              count:
                type: "integer"
                example: 1
              label:
                type: "string"
                example: "perfect_rating"
              badge_url:
                type: "string"
                example: "https://media.website.com/1554853.png"
              description:
                type: "string"
                example: "Earned on total 5-star ratings"
...
    Vehicles:
      title: "Vehicles"
      type: "object"
      properties:
        id:
          type: "string"
          description: "Object's unique identifier."
          example: "e8b085d1-fa2c-4e0f-8b2e-93c6ac991f12"
        vin:
          type: "string"
          description: "Vehicle identification number."
          example: "xuz-345"
        account:
          type: "string"
          description: "Account ID."
          example: "010db8b4-a724-47fc-c17e-733b656312a1"
        make:
          type: "string"
          description: "Make (e.g. Ford, Toyota)."
          example: "Toyota"
        model:
          type: "string"
          description: "Model (e.g. Fiesta, Prius)."
          example: "Corolla"
        year:
          type: "string"
          description: "Manufacture year (e.g. 2011)."
          example: "2011"
        identification:
          type: "string"
          description: "License plate or other id depending on the vehicle type."
          example: "HIJ8473"
        type:
          type: "string"
          description: "Vehicle type. Can be one of car, pedestrian, ebike, bicycle\
            \ or scooter."
          example: "car"
      description: "Vehicle objects represent vehicles connected to a user's workforce\
        \ account."
    Payouts:
      title: "Payouts"
      type: "object"
      properties:
        id:
          type: "string"
          description: "Object's unique identifier."
          example: "47b216e2-d334-4235-bc1e-185d15ab18d0"
        status:
          type: "string"
          description: "Payout status. The value will be one of the following completed,\
            \ scheduled or cancelled"
          example: "completed"
        type:
          type: "string"
          description: "Payout status. The value will be one of the following completed,\
            \ scheduled or cancelled"
          example: "direct_deposit"
        payout_date:
          type: "string"
          description: "Timestamp representing the payout execution date."
          example: "2020-01-27T08:26:16Z"
        currency:
          type: "string"
          description: "Currency of the payout (INR, EUR, USD)."
          example: "INR"
        gross_pay:
          type: "string"
          description: "Gross pay amount."
          example: "6.70"
        deductions:
          type: "string"
          description: "Deductions amount."
          example: "0.00"
        taxes:
          type: "string"
          description: "Taxes amount."
          example: "2.31"
        net_pay:
          type: "string"
          description: "Net pay amount."
          example: "6.70"
        bonuses:
          type: "string"
          description: "Bonuses amount."
          example: "10.70"
        commission:
          type: "string"
          description: "Commission amount"
          example: "15.32"
        overtime:
          type: "string"
          description: "Overtime pay amount."
          example: "9.02"
        reimbursements:
          type: "string"
          description: "Reimbursement amount."
          example: "20.00"
        hours:
          type: "string"
          description: "Hours spent working."
        metadata:
          type: "string"
          description: "Payout metadata."
        employer_address:
          type: "object"
          properties:
            city:
              type: "string"
              description: "City."
              example: "SINGAPORE"
            country:
              type: "string"
              description: "Country."
              example: "SINGAPORE"
            state:
              type: "string"
              description: "State."
              example: "SINGAPORE"
            line1:
              type: "string"
              description: "Address line 1."
              example: "33 IRVING PLACE"
            line2:
              type: "string"
              description: "Address line 2."
            postal_code:
              type: "string"
              description: "Postal code."
              example: "308215"
          description: "Employer Address Details"
      description: "Payouts objects contain aggregate payouts data."
...