# spring-server-generated/src/main/java/io/swagger/api/VehiclesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.21).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Vehicles;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-09-26T10:30:08.165Z[GMT]")
@Api(value = "vehicles", description = "the vehicles API")
public interface VehiclesApi {

    @ApiOperation(value = "List vehicles", nickname = "getAllVehicles", notes = "List vehicles", response = Vehicles.class, responseContainer = "List", tags={ "Vehicles", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "An object with a results property that contains an array of up to limit vehicle objects.", response = Vehicles.class, responseContainer = "List") })
    @RequestMapping(value = "/vehicles",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Vehicles>> getAllVehicles();


    @ApiOperation(value = "Retrieve a vehicle", nickname = "getVehicleById", notes = "Retrieves a vehicle object with the supplied ID.", response = Vehicles.class, tags={ "Vehicles", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Returns a vehicle object if a valid identifier was provided.", response = Vehicles.class) })
    @RequestMapping(value = "/vehicles/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Vehicles> getVehicleById(@ApiParam(value = "The identifier of the vehicle to be retrieved.",required=true) @PathVariable("id") String id
);

}